Moved things around, got rid of uneeded type usage in projection file module
